+++ app/models/scenarios/scenario_106.rb
@@ -7,7 +7,7 @@ module Scenarios
     ALLIES = ["usa"].freeze
     AXIS = ["ger"].freeze
     STATUS = "b"
-    VERSION = "0.3"
+    VERSION = "0.4"
 
     DATE = [1943, 2, 21].freeze
     LAYOUT = [15, 23, "x"].freeze
@@ -16,8 +16,8 @@ module Scenarios
       "0": {
         list: [
           :usa_leader_3_1,
-          [3, :usa_green_s],
-          :usa_m1918_bar,
+          [4, :usa_green_s],
+          [2, :usa_m1918_bar],
           :usa_m2_browning,
           [3, :usa_m3_gmc],
           [2, :foxhole],
@@ -28,11 +28,6 @@ module Scenarios
           [3, :usa_m3_lee],
         ],
       },
-      "4": {
-        list: [
-          [2, :usa_m3_lee],
-        ],
-      },
     }.freeze
 
     AXIS_UNITS = {
@@ -41,7 +36,7 @@ module Scenarios
           :ger_leader_6_2,
           [3, :ger_rifle_s],
           [2, :ger_mg_42],
-          [4, :ger_pzkpfw_iv_f2],
+          [5, :ger_pzkpfw_iv_f2],
           [2, :ger_sdkfz_251_1],
         ],
       },
@@ -101,7 +96,6 @@ module Scenarios
           allied_setup: {
             "0" => [["*", "0-14"]],
             "2" => [["*", 0]],
-            "4" => [["*", 0]],
           },
           axis_setup: { "0" => [["*", "16-22"]] },
           base_terrain: "d",